backend-springboot/python-service/check_gemini.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    print("GEMINI_API_KEY not found")
    exit(1)

genai.configure(api_key=api_key)
model = genai.GenerativeModel('gemini-1.5-flash')

try:
    response = model.generate_content("Hello, can you hear me? Respond with 'Yes' if you can.")
    print(f"Response: {response.text}")
except Exception as e:
    print(f"Error: {e}")
