backend-springboot/python-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
transformers==4.36.0
torch==2.1.2
pydub==0.25.1
gunicorn==21.2.0
tiktoken==0.5.1
tqdm==4.66.1
numpy==1.26.3
faster-whisper==0.10.0
google-generativeai>=0.5.0
python-dotenv==1.0.0